requirements.txt
asgiref==3.8.1
asttokens==2.4.1
cheap_repr==0.5.2
Django==5.1.1
djangorestframework==3.15.2
executing==2.1.0
Pygments==2.18.0
six==1.16.0
snoop==0.4.3
sqlparse==0.5.1
typing_extensions==4.12.2
